Make Bootstrap ConfigurationChecker customizable.
Also-by: Simon Bernard <code@simonbernard.eu>
